Move modal out from the card, fix search box medium size screen
