+++ app/scripts/components/exploration/components/datasets/dataset-list-item.tsx
@@ -11,6 +11,7 @@ import {
   getInteractionDataPoint,
   usePopover
 } from '../chart-popover';
+import LayerInfoModal, { LayerInfoModalData } from '../layer-info-modal';
 import {
   DatasetTrackError,
   DatasetTrackLoading
@@ -18,6 +19,7 @@ import {
 import { DatasetChart } from './dataset-chart';
 import { getBlockBoundaries, lumpBlocks } from './block-utils';
 import DataLayerCard from './data-layer-card';
+import { findDatasetAttribute } from '$components/exploration/data-utils';
 import {
   TimelineDatasetStatus,
   TimelineDatasetSuccess
@@ -99,7 +101,7 @@ export function DatasetListItem(props: DatasetListItemProps) {
   const { isAnalyzing, runAnalysis } = useAnalysisController();
 
   const [isVisible, setVisible] = useTimelineDatasetVisibility(datasetAtom);
-
+  const [modalLayerInfo, setModalLayerInfo] = React.useState<LayerInfoModalData>();
   const queryClient = useQueryClient();
 
   const retryDatasetMetadata = useCallback(() => {
@@ -111,6 +113,19 @@ export function DatasetListItem(props: DatasetListItemProps) {
       { throwOnError: false }
     );
   }, [queryClient, datasetId]);
+
+  const onClickLayerInfo = useCallback(() => {
+    const parentInfoDesc = findDatasetAttribute({datasetId: dataset.data.parentDataset.id, attr: 'infoDescription'});
+    const data: LayerInfoModalData = {
+      name: dataset.data.name,
+      info: dataset.data.info,
+      parentData: {
+        ...dataset.data.parentDataset,
+        infoDescription: parentInfoDesc
+      }
+    };
+    setModalLayerInfo(data);
+  }, [dataset]);
 
   const controls = useDragControls();
 
@@ -184,8 +199,15 @@ export function DatasetListItem(props: DatasetListItemProps) {
         <DatasetHeader>
           <DatasetHeaderInner>
             <div style={{width: '100%'}} onPointerDown={onDragging}>
-              <DataLayerCard dataset={dataset} datasetAtom={datasetAtom} isVisible={isVisible} setVisible={setVisible} datasetLegend={datasetLegend} />
+              <DataLayerCard dataset={dataset} datasetAtom={datasetAtom} isVisible={isVisible} setVisible={setVisible} datasetLegend={datasetLegend} onClickLayerInfo={onClickLayerInfo} />
             </div>
+            {modalLayerInfo && (
+              <LayerInfoModal
+                revealed={!!modalLayerInfo}
+                close={() => setModalLayerInfo(undefined)}
+                layerData={modalLayerInfo}
+              />
+            )}
           </DatasetHeaderInner>
         </DatasetHeader>
         <DatasetData>
